feat(projects): add Syncfusion scripts and header
Added a new script to load the Syncfusion library and defined a function to open a new page when clicking on the element with ID `go-project`. Also included an image and a header displaying the value of `@Model.Header.Value`.

In the footer, adjusted the class order for better readability and added a link to the imprint page under the copyright section. Additionally, modified the class order in the mobile language selection section to optimize structure without affecting functionality.